Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,10 +3,10 @@
 int main()
 {
     int x;
-    float area=0;
+    double area=0;
     scanf("%d",&x);
     printf("");
     area= x*x*3.14;
-    printf("Area: %.2f",area);
+    printf("Area: %.2lf",area);
     return 0;
 }
